pyproject.toml
[tool.poetry]
name = "infra"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
requests = "^2.31.0"
gitpython = "^3.1.43"
python-gitlab = "^3.0.0"
ruamel-yaml = ">=0.17.21,<0.18.0"
demisto-sdk = "1.31.2"
coloredlogs = "^15.0.1"
pre-commit = "^3.7.0"
pandas = "^2.2.2"
google = "^3.0.0"
google-cloud-bigquery = "^3.22.0"
jira = "^3.8.0"
timeout-decorator = "^0.5.0"
requests-mock = "^1.12.1"
tqdm = "^4.66.4"
typer = "^0.9.0"
jinja2 = "^3.1.3"
pendulum = "^2.1.2"
google-cloud-compute = "^1.8.0"
google-api-python-client = "^2.64.0"
google-auth-httplib2 = "^0.1.0"
google-auth-oauthlib = "^0.5.3"
google-cloud-firestore = "^2.3.4"
google-cloud-logging = "^3.1.2"
google-cloud-pubsub = "^2.18.3"
google-cloud-secret-manager = "^2.11.1"
google-cloud-storage = "^2.9.0"
retry = "^0.9.2"
Faker = "^15.3.4"
sendgrid = "^6.11.0"
pygithub = "^2.3.0"
humanize = "^4.9.0"
xmltodict = "^0.13.0"
beautifulsoup4 = "^4.11.1"
deepdiff = "^6.3.0"
filelock = "^3.10.0"
glom = "^23.3.0"
mergedeep = "^1.3.4"
more-itertools = "^9.1.0"
paramiko = "^2.12.0"
polling2 = "^0.5.0"
pydantic = "^1.10.8"
PyYAML = "^6.0"
statsd = "^3.3.0"
termcolor = "^1.1.0"


[tool.mypy]
python_version = "3.10"
check_untyped_defs = true
ignore_missing_imports = true
namespace_packages = true
show_error_codes = true
follow_imports = "silent"
pretty = true
allow_redefinition = true
exclude = [
    "test_data",
    "tests_data",
    ".venv",
    ".*_test.py$",
    "test_.*.py",
    "*infrastructure_tests*",
    "BlacklistValidation*",
    "SecretActions*",
    "artifacts*"
]

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
pytest-check="1.0.10"
pytest-xdist="3.2.1"
pytest-mock = "^3.14.0"
mock-open = "^1.4.0"
mock = "^5.1.0"

[tool.poetry.group.linters.dependencies]
mypy = "^0.982"
ruff = "^0.0.264"


[tool.poetry.group.testing]
optional = true

[tool.poetry.group.testing.dependencies]
pytest-cov = "^5.0.0"

[tool.ruff]
ignore = [
    "F403", # undefined-local-with-import-star
    "F405", # undefined-local-with-import-star-usage
    "PLC1901", #compare-to-empty-string
]
select =['E','F','PLC','PLE']
extend-exclude = ["Tests/scripts/infrastructure_tests/tests_data", "Utils/test_upload_flow"]
line-length = 130

[tool.pytest.ini_options]
addopts = """
    --disable-warnings
    --color=yes
    -v
    --ignore='Tests/test_integration.py'
    --ignore='Tests/tests_e2e'
    --ignore='BlacklistValidation/test_data/Packs/Grafana/Integrations/Grafana/Grafana_test.py'
    --junitxml=report.xml
    """
python_files = "test_*.py *_test.py"

[tool.coverage.run]
relative_files = true
omit = ["*/test_data/*", "*_test.py", "conftest.py", "**/conftest.py"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
